requirements.txt
torch_nightly==1.1.0.dev20190502
tqdm==4.40.2
h5py==2.10.0
tensorflow==2.2.0
nltk==3.4.5
numpy==1.17.4
flair==0.4.5
transformers==4.5.1
python_dateutil==2.8.2
scikit_learn==1.0
torch==1.10.0
